[JD-9] Style: 코드 컨벤션에 맞게 index.tsx 로 통일
