+++ src/components/LikeButton/index.tsx
@@ -1,12 +1,13 @@
-import React from "react";
-import { IoPaw, IoPawOutline } from "react-icons/io5";
+import React from 'react';
+
+import { IoPaw, IoPawOutline } from 'react-icons/io5';
 
 interface LikeProps {
   like: boolean;
   onClick: () => void;
 }
 
-const LikeButton = ({ like, onClick }: LikeProps) => {
+const LikeButton: React.FC<LikeProps> = ({ like, onClick }) => {
   return <div onClick={onClick}>{like ? <IoPaw /> : <IoPawOutline />}</div>;
 };
