+++ src/components/Loading/index.tsx
@@ -1,9 +1,12 @@
-import LoadingSpinner from "./LoadingSpinner";
-import { useIsFetching } from "@tanstack/react-query";
+import React from 'react';
+
+import { useIsFetching } from '@tanstack/react-query';
+
+import LoadingSpinner from './LoadingSpinner';
 
 export function Loading() {
   const isFetching = useIsFetching();
-  const display = isFetching ? "inherit" : "none";
+  const display = isFetching ? 'inherit' : 'none';
 
   return (
     <div style={{ display: display }}>
